benchmarks/original/video-analytics/video-decoder/requirements.txt
parliament-functions==0.1.0
environs==9.3.2
pycapnp==1.3.0
opencv-python-headless
requests
tracing
